Remove wrapper around error message in response
This is done for now to provide a consistent feel to the API.
Ultimately I'd like to wrap the message to maybe give more information
in the future. However, Ring rejects some requests entirely (e.g
malformed json body). As a result I can't enforce anything there.
Something to look into in the future.